Eliminate the duplicate model directory -- move model into linkml_model.
We _tried_ to make model appear in linkml_model via setup.cfg, but it was
infuriatingly difficult.  It ended up being easier to work with the gitignore instead.

All of the contents of linkml_model are included in the pypi binary distribution.  model gets committed and everything else is generated.