Use versioned documentation deployment
These changes use mike (https://github.com/jimporter/mike) to have multiple versions of
the documentation deployed. Changes to the main branch will be deployed as a "dev" version.
Each release is deployed as a "MAJOR.MINOR" version -- that is, releasing 1.5.3 will replace
the 1.5.2 docs. But releasing 1.6.0 will not affect the 1.5.3 docs. The latest version will
also be accessible via the "latest" alias.