fix: decouple 'curie' uri from 'uri' uri
The URIs assigned to the different types identify the types themselves,
as a consequence two different types with the same URI would be expected
to be synonyms (different names with the same meaning).

This is the case for the types CURIE and URIorCURIE:
1. CURIE had the URI 'xsd:anyURI' (the same as for URIs).
2. URIorCURIE had the URI 'xsd:string' (the same as for strings).

This patch assign both CURIE and URIorCURIE their own unique URIs to
differentiate the different types from each other.

Signed-off-by: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>